+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Check for overlapping microhaps from different sources during the database build (#112).
 - Over 1,000 marker definitions from four whole-genome screens from Yu et al 2022 (see #115, #118).
 - New `nomenclature` module for parsing and validating microhap identifiers (see #119).
+- New `summarize` subcommand for summarizing MicroHapDB database contents (see #120).
 
 ### Changed
 - Overhaul of the marker table structure and database build procedure (see #106, #111, #112).